requirements.txt
attrs==21.4.0
autopep8==1.6.0
boto3==1.21.8
botocore==1.24.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.8
Werkzeug==2.0.3
